+++ server/src/main/java/com/sap/vcs/server/repository/DocumentVersionRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DocumentVersionRepository extends JpaRepository<DocumentVersion, Long> {
+public interface DocumentVersionRepository extends JpaRepository<DocumentVersion, Integer> {
 }
